created lessons endpoint, made migrations
