+++ backend/backend/settings.py
@@ -43,6 +43,7 @@ INSTALLED_APPS = [
     "django.contrib.staticfiles",
     'rest_framework',
     "music.apps.MusicConfig",
+    "lessons.apps.LessonsConfig",
     "account.apps.AccountConfig",
 ]
