+++ backend/lessons/admin.py
@@ -1,0 +1,4 @@
+from django.contrib import admin
+from .models import Lesson
+
+admin.site.register(Lesson)
